Feat/leo/place review

+++ src/main/java/DC_square/spring/repository/place/PlaceDetailRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface PlaceDetailRepository extends JpaRepository<PlaceDetail, Integer> {
+public interface PlaceDetailRepository extends JpaRepository<PlaceDetail, Long> {
     Optional<PlaceDetail> findByPlace(Place place);
-    Optional<PlaceDetail> findByPlaceId(Integer placeId);
+    Optional<PlaceDetail> findByPlaceId(Long placeId);
 }
